pyproject.toml
[tool.poetry]
name = "devbot"
version = "0.1.0"
description = ""
authors = ["tolatolatop <tolatolatop@gmail.com>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pyyaml = "^6.0.1"
langchain = "^0.0.334"
python-dotenv = "^1.0.0"
openai = "^1.2.3"
gitpython = "^3.1.40"
pygithub = "^2.1.1"
beautifulsoup4 = "^4.12.2"
html2text = "^2020.1.16"
unstructured = "^0.11.0"
pygitee = {git = "https://github.com/tolatolatop/pygitee.git"}
redis = "^5.0.1"


[tool.poetry.group.dev.dependencies]
wheel = "^0.41.3"
watchdog = "^3.0.0"
flake8 = "^6.1.0"
tox = "^4.11.3"
coverage = "^7.3.2"
sphinx = "^7.2.6"
twine = "^4.0.2"
pytest = "^7.4.3"
black = "^23.11.0"
httpx = "^0.25.1"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79